feat: Cooldown data component refactor: moved AbilityHolder into ability subpackage refactor: tweaked AbilitySlots serialization setup slightly for encapsulation
