+++ src/main/java/com/wanderersoftherift/wotr/init/WotrDataComponentType.java
@@ -5,6 +5,7 @@ import com.wanderersoftherift.wotr.WanderersOfTheRift;
 import com.wanderersoftherift.wotr.abilities.AbstractAbility;
 import com.wanderersoftherift.wotr.abilities.upgrade.AbilityUpgradePool;
 import com.wanderersoftherift.wotr.item.LootBox;
+import com.wanderersoftherift.wotr.item.ability.Cooldown;
 import com.wanderersoftherift.wotr.item.currency.CurrencyProvider;
 import com.wanderersoftherift.wotr.item.implicit.GearImplicits;
 import com.wanderersoftherift.wotr.item.riftkey.RiftConfig;
@@ -41,8 +42,11 @@ public class WotrDataComponentType {
             "item_rift_tier", Codec.INT, null);
     public static final DeferredHolder<DataComponentType<?>, DataComponentType<LootBox>> LOOT_BOX = register("loot_box",
             LootBox.CODEC, null);
+
     public static final DeferredHolder<DataComponentType<?>, DataComponentType<Holder<AbstractAbility>>> ABILITY = register(
             "ability", AbstractAbility.CODEC, AbstractAbility.STREAM_CODEC);
+    public static final DeferredHolder<DataComponentType<?>, DataComponentType<Cooldown>> COOLDOWN = register(
+            "cooldown", Cooldown.CODEC, null);
     public static final DeferredHolder<DataComponentType<?>, DataComponentType<AbilityUpgradePool>> ABILITY_UPGRADE_POOL = register(
             "ability_upgrade_pool", AbilityUpgradePool.CODEC, AbilityUpgradePool.STREAM_CODEC);
